# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QuotaEntry": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining"
          },
          "tracked": {
            "title": "Tracked",
            "type": "boolean"
          },
          "used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Used"
          }
        },
        "required": [
          "key",
          "label",
          "period",
          "limit",
          "used",
          "remaining",
          "available",
          "tracked"
        ],
        "title": "QuotaEntry",
        "type": "object"
      },
      "QuotaSummaryEntry": {
        "properties": {
          "available": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available"
          },
          "resource": {
            "title": "Resource",
            "type": "string"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "required": [
          "resource",
          "total",
          "available"
        ],
        "title": "QuotaSummaryEntry",
        "type": "object"
      },
      "QuotaSummaryResponse": {
        "properties": {
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolvedQuota"
              },
              {
                "type": "null"
              }
            ]
          },
          "quotas": {
            "items": {
              "$ref": "#/components/schemas/QuotaSummaryEntry"
            },
            "title": "Quotas",
            "type": "array"
          },
          "tier_name": {
            "title": "Tier Name",
            "type": "string"
          }
        },
        "required": [
          "tier_name"
        ],
        "title": "QuotaSummaryResponse",
        "type": "object"
      },
...
      "ResolvedQuota": {
        "properties": {
          "availability": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Availability",
            "type": "object"
          },
          "highest_lesson_depth": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highest Lesson Depth"
          },
          "max_file_upload_kb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max File Upload Kb"
          },
          "max_sections_per_lesson": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Sections Per Lesson"
          },
          "quota_entries": {
            "items": {
              "$ref": "#/components/schemas/QuotaEntry"
            },
            "title": "Quota Entries",
            "type": "array"
          },
          "remaining_file_uploads": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining File Uploads"
          },
          "remaining_image_uploads": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Image Uploads"
          },
          "remaining_research": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Research"
          },
          "remaining_sections": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Sections"
          },
          "tier_name": {
            "title": "Tier Name",
            "type": "string"
          },
          "total_file_uploads": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total File Uploads"
          },
          "total_image_uploads": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Image Uploads"
          },
          "total_research": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Research"
          },
          "total_sections": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Sections"
          },
          "tutor_mode_enabled": {
            "title": "Tutor Mode Enabled",
            "type": "boolean"
          },
          "tutor_voice_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tutor Voice Tier"
          }
        },
        "required": [
          "tier_name",
          "max_file_upload_kb",
          "highest_lesson_depth",
          "max_sections_per_lesson",
          "total_file_uploads",
          "total_image_uploads",
          "total_sections",
          "total_research",
          "remaining_file_uploads",
          "remaining_image_uploads",
          "remaining_sections",
          "remaining_research",
          "tutor_mode_enabled",
          "tutor_voice_tier"
        ],
        "title": "ResolvedQuota",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me": {
      "get": {
        "description": "Get the current user's profile and onboarding status.",
        "operationId": "get_my_profile_api_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get My Profile Api Me Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Profile",
        "tags": [
          "onboarding"
        ]
      }
    },
...
    "/api/user/me": {
      "get": {
        "description": "Get the current user's profile.",
        "operationId": "get_my_profile_api_user_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get My Profile Api User Me Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Profile",
        "tags": [
          "users"
        ]
      }
    },
    "/api/user/me/features": {
      "get": {
        "description": "Get effective feature flags, runtime config, and permission hints for the current user.",
        "operationId": "get_my_features_api_user_me_features_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get My Features Api User Me Features Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Features",
        "tags": [
          "users"
        ]
      }
    },
    "/api/user/me/quota": {
      "get": {
        "description": "Get the current user's quota and subscription tier.",
        "operationId": "get_my_quota_api_user_me_quota_get",
        "parameters": [
          {
            "in": "query",
            "name": "details",
            "required": false,
            "schema": {
              "default": false,
              "title": "Details",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Quota",
        "tags": [
          "users"
        ]
      }
    },
...
    "/v1/tutor/audio/{audio_id}/content": {
      "get": {
        "description": "Stream audio content.",
        "operationId": "get_audio_content_v1_tutor_audio__audio_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "audio_id",
            "required": true,
            "schema": {
              "title": "Audio Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Audio Content",
        "tags": [
          "tutor"
        ]
      }
    },
    "/v1/tutor/job/{job_id}/audios": {
      "get": {
        "description": "Retrieve list of generated audios for a job.",
        "operationId": "get_job_audios_v1_tutor_job__job_id__audios_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job Audios V1 Tutor Job  Job Id  Audios Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Job Audios",
        "tags": [
          "tutor"
        ]
      }
    },
...
  }
}